+++ lib/pubsub_service.js
@@ -1,5 +1,5 @@
-var Logger = require('./logger');
-var l = Logger();
+//var Logger = require('./logger');
+//var l = Logger();
 
 var subscribedTo = [];
 var response;
@@ -36,7 +36,7 @@ exports.publish = function(name, data) {
 exports.subscribe = function(res, name) {
   response = res;
   if (subscribedTo.indexOf(name) === -1) {
-    l.emit('log', 'fog-runtime', 'Created subscription to stream '+name);
+//    l.emit('log', 'fog-runtime', 'Created subscription to stream '+name);
     subscribedTo.push(name);
   }
 };
