+++ lib/scientist.js
@@ -1,4 +1,3 @@
-var MachineConfig = require('./machine_config');
 
 exports.configure = function(/* constructor, ...constructorArgs */) {
   var args = Array.prototype.slice.call(arguments);
@@ -14,9 +13,7 @@ exports.configure = function(/* constructor, ...constructorArgs */) {
     machine = constructor;
   }
 
-  var config = MachineConfig.create(machine);
-
-  machine.init(config);
+  machine.init(machine);
 
   return machine;
 };
@@ -39,9 +36,8 @@ exports.create = function(/* constructor, ...constructorArgs */) {
 };
 
 exports.init = function(machine) {
-  var config = MachineConfig.create(machine);
 
-  machine.init(config);
+  machine.init(machine);
 
   return machine;
 }
